requirements.txt
google-api-python-client
google-auth-httplib2
oauth2client
discord.py
